Sliming down
Sliming down the framework to gain performance (and replace ORM)
